move Page to main types
