+++ .changeset/stupid-beds-prove.md
@@ -2,4 +2,4 @@
 '@sveltejs/kit': patch
 ---
 
-add new `Page` type in `$app/stores`
+add new `Page` type
